Add support for poe_tasks.toml|yaml|json files
