+++ poethepoet/config/__init__.py
@@ -1,0 +1,4 @@
+from .config import PoeConfig
+from .partition import KNOWN_SHELL_INTERPRETERS, ConfigPartition
+
+__all__ = ["PoeConfig", "ConfigPartition", "KNOWN_SHELL_INTERPRETERS"]
