+++ pyproject.toml
@@ -16,21 +16,23 @@ python = ">=3.8"
 pastel = "^0.2.1"
 tomli  = ">=1.2.2"
 poetry = {version = "^1.0", allow-prereleases = true, optional = true}
+pyyaml = "^6.0.2"
 
 [tool.poetry.group.ci.dependencies]
-black      = "^23.3.0"
-mypy       = "^1.1.1"
-pytest     = "^7.1.2"
-pytest-cov = "^3.0.0"
-rstcheck   = { version = "^6.1.2", python = "<4" }
-ruff       = "^0.0.291"
+black        = "^23.3.0"
+mypy         = "^1.1.1"
+pytest       = "^7.1.2"
+pytest-cov   = "^3.0.0"
+rstcheck     = { version = "^6.1.2", python = "<4" }
+ruff         = "^0.0.291"
+types-pyyaml = "^6.0.12.20240808"
 
 virtualenv = "^20.14.1"
 
 poe_test_helpers = { path = "./tests/fixtures/packages/poe_test_helpers" }
 
 [tool.poetry.group.dev.dependencies]
-bpython    = "^0.24"
+bpython      = "^0.24"
 
 [tool.poetry.group.docs.dependencies]
 furo                = "^2023.3.27"
