+++ tests/fixtures/poe_tasks_file_project/sub1/poe_tasks.json
@@ -1,0 +1,9 @@
+{
+  "tool": {
+    "poe": {
+      "tasks": {
+        "sub1_poetasks_json": "poe_test_echo sub1_poetasks_json"
+      }
+    }
+  }
+}
