+++ tests/fixtures/poe_tasks_file_project/sub1/poe_tasks.toml
@@ -1,0 +1,3 @@
+
+[tool.poe.tasks.sub1_poetasks_toml]
+cmd = "poe_test_echo sub1_poetasks_toml"
