+++ tests/fixtures/poe_tasks_file_project/sub2/poe_tasks.yaml
@@ -1,0 +1,2 @@
+tasks:
+  sub2_poetasks_yaml: "poe_test_echo sub2_poetasks_yaml"
