+++ tests/fixtures/poe_tasks_file_project/sub2/pyproject.toml
@@ -1,0 +1,1 @@
+# this pyproject.toml doesn't contain any poe config
