+++ tests/fixtures/poe_tasks_file_project/sub3/poe_tasks.json
@@ -1,0 +1,6 @@
+{
+  "include": "../sub2/poe_tasks.yaml",
+  "tasks": {
+    "sub3_poetasks_json": "poe_test_echo sub3_poetasks_json"
+  }
+}
